feat(estimate-builder): pinned bottom totals bar (TotalsBar) replaces floating panel (#545)
Retire the floating TotalsPanel card in favor of a full-width TotalsBar
pinned to the bottom of the builder via BuilderLayout's totals slot (#543).

- Subtotal, Markup, Discount, Adjusted subtotal, Tax, and grand Total all
  render in the bar.
- Markup, Discount, and Tax are editable inline; fixed-amount Markup/Discount
  use the $-prefixed MoneyInput (#542), Tax keeps a % box.
- A negative grand total shows the warning.
- Hidden in Template mode.
- Compact tap-to-expand on phone; always inline on desktop.
- Works in both estimate (total) and invoice (total_amount) modes.

Wired into the estimate and invoice branches; the two floating <TotalsPanel>
siblings and totals-panel.tsx/.test.tsx are deleted.

12 TotalsBar tests cover the acceptance criteria. tsc clean; lint clean on
touched files.

Closes #545
